Update debian changelog.
Declare comerr-dev as replacing << e2fslibs-dev 1.33-2, to avoid
        errors when upgrading to the new versions of comerr-dev and
        e2fslibs-dev

Declare init_error_table as taking a long for the second argument.

+++ lib/et/init_et.c
@@ -31,7 +31,7 @@ struct foobar {
 
 extern struct et_list * _et_list;
 
-int init_error_table(const char * const *msgs, int base, int count)
+int init_error_table(const char * const *msgs, long base, int count)
 {
     struct foobar * new_et;
